style: refactor globals styles
