+++ src/styles/globals.ts
@@ -1,5 +1,4 @@
-import { createGlobalStyle, css, DefaultTheme } from 'styled-components';
-import { COLOR } from './colors';
+import { createGlobalStyle } from 'styled-components';
 
 export const GlobalStyles = createGlobalStyle`
   *,
@@ -17,7 +16,7 @@ export const GlobalStyles = createGlobalStyle`
     text-rendering: optimizelegibility;
     -webkit-font-smoothing: antialiased;
     -moz-osx-font-smoothing: grayscale;
-    font-size: 18px;
+    font-size: 16px;
     line-height: 1.5;
     transition: all 1s ease;
     -ms-overflow-style: none;  /* Internet Explorer 10+ */
@@ -37,37 +36,3 @@ export const GlobalStyles = createGlobalStyle`
     }
   }
 `;
-
-export const Heading = css`
-  margin-bottom: 24px;
-  background: ${COLOR.linearGradient};
-  background-clip: text;
-  -webkit-background-clip: text;
-  -webkit-text-fill-color: transparent;
-  font-size: 32px;
-  font-weight: 600;
-  line-height: 1.1;
-  @media screen and (min-width: 480px) {
-    font-size: 48px;
-  }
-`;
-
-export const lightTheme: DefaultTheme = {
-  body: COLOR.lightBackground,
-  text: COLOR.lightText,
-  textSecondary: COLOR.lightTextSecondary,
-  primary: COLOR.lightPrimary,
-  navbarBg: COLOR.lightNavbar,
-  navbarLink: COLOR.lightNavbarLink,
-  footer: COLOR.lightFooter,
-};
-
-export const darkTheme: DefaultTheme = {
-  body: COLOR.darkBackground,
-  text: COLOR.darkText,
-  textSecondary: COLOR.darkTextSecondary,
-  primary: COLOR.darkPrimary,
-  navbarBg: COLOR.darkNavbar,
-  navbarLink: COLOR.darkNavbarLink,
-  footer: COLOR.darkFooter,
-};
